DailyData updates - share the write function
