+++ src/op_analytics/datasources/ethereum_optimism/dataaccess.py
@@ -1,10 +1,5 @@
-import polars as pl
-
 from op_analytics.coreutils.logger import structlog
-from op_analytics.coreutils.partitioned.dailydata import (
-    write_daily_data,
-    DailyDataset,
-)
+from op_analytics.coreutils.partitioned.dailydata import DailyDataset
 
 log = structlog.get_logger()
 
@@ -14,14 +9,3 @@ class EthereumOptimism(DailyDataset):
 
     # Superchain Token List
     SUPERCHAIN_TOKEN_LIST = "superchain_token_list_v1"
-
-    def write(
-        self,
-        dataframe: pl.DataFrame,
-        sort_by: list[str] | None = None,
-    ):
-        return write_daily_data(
-            root_path=self.root_path,
-            dataframe=dataframe,
-            sort_by=sort_by,
-        )
