+++ src/op_analytics/datasources/github/dataaccess.py
@@ -1,10 +1,5 @@
-import polars as pl
-
 from op_analytics.coreutils.logger import structlog
-from op_analytics.coreutils.partitioned.dailydata import (
-    write_daily_data,
-    DailyDataset,
-)
+from op_analytics.coreutils.partitioned.dailydata import DailyDataset
 
 log = structlog.get_logger()
 
@@ -24,14 +19,3 @@ class Github(DailyDataset):
     PRS = "github_prs_v1"
     PR_COMMENTS = "github_pr_comments_v1"
     PR_REVIEWS = "github_pr_reviews_v2"
-
-    def write(self, dataframe: pl.DataFrame, sort_by: list[str] | None = None):
-        """
-        Write DataFrame to GCS (daily partitioned) using the root_path
-        for the given dataset constant (e.g. COMMITS, ISSUES, etc.).
-        """
-        return write_daily_data(
-            root_path=self.root_path,
-            dataframe=dataframe,
-            sort_by=sort_by,
-        )
