+++ tests/op_analytics/cli/pulls/defillama/test_defillama_protocols.py
@@ -414,7 +414,7 @@ def test_construct_urls():
 
 @patch("op_analytics.datasources.defillama.protocols.get_data")
 @patch("op_analytics.datasources.defillama.protocols.now_date")
-@patch("op_analytics.coreutils.partitioned.dailydata.PartitionedWriteManager.write")
+@patch("op_analytics.coreutils.partitioned.dailydatawrite.PartitionedWriteManager.write")
 def test_pull_single_protocol_tvl(
     mock_write,
     mock_now_date,
@@ -507,7 +507,7 @@ def test_pull_single_protocol_tvl(
 
 @patch("op_analytics.datasources.defillama.protocols.get_data")
 @patch("op_analytics.datasources.defillama.protocols.now_date")
-@patch("op_analytics.coreutils.partitioned.dailydata.PartitionedWriteManager.write")
+@patch("op_analytics.coreutils.partitioned.dailydatawrite.PartitionedWriteManager.write")
 def test_pull_all_protocol_tvl(
     mock_write,
     mock_now_date,
